Move HandshakeProtocol to other file
